+++ src/heat.js
@@ -199,9 +199,9 @@
             var days = createElement( "div", "days" );
             map.appendChild( days );
 
-            if ( !bindingOptions.showMonthNames ) {
+            if ( !bindingOptions.showMonthNames || bindingOptions.placeMonthNamesOnTheBottom ) {
                 days.style.paddingTop = "0px";
-                days.style.marginTop = "-5px";
+                days.style.marginTop = !bindingOptions.placeMonthNamesOnTheBottom ? "-5px":  "-2px";
             }
     
             for ( var dayNameIndex = 0; dayNameIndex < 7; dayNameIndex++ ) {
@@ -225,10 +225,10 @@
                 var month = createElement( "div", "month" );
                 months.appendChild( month );
     
-                if ( bindingOptions.showMonthNames ) {
-                    var monthName = createElement( "div", "month-name" );
-                    monthName.innerHTML = _configuration.monthNames[ monthIndex ];
-                    month.appendChild( monthName );
+                if ( bindingOptions.showMonthNames && !bindingOptions.placeMonthNamesOnTheBottom ) {
+                    var monthNameTop = createElement( "div", "month-name" );
+                    monthNameTop.innerHTML = _configuration.monthNames[ monthIndex ];
+                    month.appendChild( monthNameTop );
                 }
     
                 var dayColumns = createElement( "div", "day-columns" );
@@ -268,6 +268,12 @@
                     }
 
                     actualDay++;
+                }
+
+                if ( bindingOptions.showMonthNames && bindingOptions.placeMonthNamesOnTheBottom ) {
+                    var monthNameBottom = createElement( "div", "month-name-bottom" );
+                    monthNameBottom.innerHTML = _configuration.monthNames[ monthIndex ];
+                    month.appendChild( monthNameBottom );
                 }
 
                 if ( monthAdded && isDefined( _elements_Day_Width ) ) {
@@ -562,6 +568,7 @@
         options.showMonthNames = getDefaultBoolean( options.showMonthNames, true );
         options.showExportButton = getDefaultBoolean( options.showExportButton, false );
         options.mapTogglesEnabled = getDefaultBoolean( options.mapTogglesEnabled, true );
+        options.placeMonthNamesOnTheBottom = getDefaultBoolean( options.placeMonthNamesOnTheBottom, false );
 
         if ( isInvalidOptionArray( options.monthsToShow ) ) {
             options.monthsToShow = [ 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12 ];