+++ docs/CHANGE_LOG.md
@@ -2,6 +2,9 @@
 
 ## Version 0.9.0:
 
+#### **Binding Options:**
+- Added a new binding option called "placeMonthNamesOnTheBottom", which states if the month names should be placed at the bottom (defaults to false).
+
 #### **General Improvements:**
 - When the binding option "showMonthDayGaps" is set to true, gaps are now always shown between the months (even if the first day is a monday).